--- apps/client/src/components/Pages/recherche/NotDeployedBanner/index.ts
@@ -1,3 +1,0 @@
-import NotDeployedBanner from "./NotDeployedBanner";
-
-export default NotDeployedBanner;
