+++ apps/client/src/locales/ar/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "مواضيع الاندماج",
     "positionButton": "استخدام موقعي",
     "notDeployedTitle": "يتم تخديم هذا الاقليم تباعاً !",
-    "notDeployedText": "Réfugiés.info هي منصة تشاركية, تجد فيها معلومات تضعها الهيئات الاجتماعية و المساعدة الاجتماعيين و المتطوعين, أنت أيضاً, يُمكنك المشاركة في اغناء المنصّة !",
     "notDeployedWriteLink": "تحرير صفحة معلومات",
     "notDeployedOkLink": "فهمت",
     "france": "في كل فرنسا",
