+++ apps/client/src/locales/en/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "Themes of integration",
     "positionButton": "Use my position",
     "notDeployedTitle": ": this department is currently being deployed!",
-    "notDeployedText": "Réfugiés.info is a collaborative site, fed directly by associations, volunteers and social workers. Help us enrich it!",
     "notDeployedWriteLink": "Write a sheet",
     "notDeployedOkLink": "I understand",
     "france": "all over France",
