+++ apps/client/src/locales/fa/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "موضوعات ادغام",
     "positionButton": "از موقعیت من استفاده کنید",
     "notDeployedTitle": ": این دپارتمان در حال گسترش است",
-    "notDeployedText": "Réfugiés.info پلتفرم ای است که مستقیماً توسط انجمن ها، داوطلبان و مددکاران اجتماعی پشتیبانی می شود. به ما کمک کنید این برنامه را غنی سازیم!",
     "notDeployedWriteLink": "برگه بنویسید",
     "notDeployedOkLink": "فهمیده شد",
     "france": "کل فرانسه",
