+++ apps/client/src/locales/fr/common.json
@@ -92,7 +92,7 @@
     "titleThemes": "Les thématiques de l'intégration",
     "positionButton": "Utiliser ma position",
     "notDeployedTitle": " : ce département est en cours de déploiement !",
-    "notDeployedText": "Réfugiés.info est collaboratif, alimenté directement par les acteurs associatifs, les bénévoles et les intervenants sociaux. Aidez-nous à l’enrichir ! ",
+    "notDeployedText": "Le référencement des actions locales débute dans le département – {{department}}. Votre recherche peut aboutir à peu de résultats.",
     "notDeployedWriteLink": "Rédiger une fiche",
     "notDeployedOkLink": "J'ai compris",
     "france": "toute la France",
