+++ apps/client/src/locales/ps/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "د ادغام موضوعات",
     "positionButton": "زما موقعیت وکاروئ",
     "notDeployedTitle": "دا څانګه په حال دې اضافه کولو کې ده !",
-    "notDeployedText": "Réfugiés.info یو ملاتړ کونکی پلټفارم دی، چې په مستقیم ډول د اتحادیو، رضاکارانو او ټولنیزو کارکونکو لخوا پکي مالوماتي فایلونه ځای په ځای کیږي. په بډایه کولو کې یې له مونږ مرسته وکړئ!",
     "notDeployedWriteLink": "یو فایل ولیکئ",
     "notDeployedOkLink": "زه پوه شوم",
     "france": "ټوله فرانسه",
