+++ apps/client/src/locales/ru/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "Темы интеграции",
     "positionButton": "Используйте мою позицию",
     "notDeployedTitle": ": этот департамент находится в процессе подготовки",
-    "notDeployedText": "Réfugiés.info - это совместный проект, который напрямую поддерживается ассоциациями, волонтерами и социальными работниками. Помогите нам дополнить его!",
     "notDeployedWriteLink": "Создать страницу",
     "notDeployedOkLink": "Я понял",
     "france": "Вся Франция",
