+++ apps/client/src/locales/ti/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "ን ምውህሃድ ዝሕግዙ ዛዕባታት",
     "positionButton": "ኣቀማምጣይ ተጠቐመሉ",
     "notDeployedTitle": "ኣብዚ ዞባ'ዚ፣ ጌና ኣገልግሎት ኣትኣተዎን",
-    "notDeployedText": "ረፉጂ ኣንፎ፣ ንናይ ማሕበራት: ወለንተኛትን: ወሃብቲ ኣገልግሎት ትካላትን ሓበሬታ ብቐጥታ ብምሕባር ምሕጋዙ ይቕጽል",
     "notDeployedWriteLink": "እቲ ዝድለ ፋይል ጽሓፍዎ",
     "notDeployedOkLink": "ተረዲኡኒ  ወይ ርዱእ",
     "france": "ኣብ ኵሉ ሃገረ ፈረንሳይ",
