+++ apps/client/src/locales/uk/common.json
@@ -89,7 +89,6 @@
     "titleThemes": "Тематики інтеграції",
     "positionButton": "Використовуйте мою позицію",
     "notDeployedTitle": ": інформація щодо цього департаменту у стані обробки!",
-    "notDeployedText": "Réfugiés.info — це спільний проєкт, який об'єднує інформацію від асоціацій, волонтерів та соціальних працівників. Допоможіть нам збагатити його актуальною інформацією!",
     "notDeployedWriteLink": "Створити запис",
     "notDeployedOkLink": "Зрозуміло",
     "france": "у всій Франції",
